Replaced IDX cache actions / triggers with composeDB ones
