+++ hasura/metadata/cron_triggers.yaml
@@ -1,15 +1,3 @@
-- name: queueExpiredIDXCacheEntries
-  webhook: '{{ACTION_BASE_ENDPOINT}}/idxCache/checkExpired'
-  schedule: 0 3 * * *
-  include_in_metadata: true
-  payload: {}
-  retry_conf:
-    num_retries: 0
-    timeout_seconds: 600
-    tolerance_seconds: 21600
-    retry_interval_seconds: 10
-  comment: Checks for cache entries more than four days old and queues them to be
-    recached.
 - name: syncAllGuildDiscordMembers
   webhook: '{{ACTION_BASE_ENDPOINT}}/syncAllGuildDiscordMembers'
   schedule: 31 5 * * *
@@ -26,3 +14,10 @@
     tolerance_seconds: 21600
     retry_interval_seconds: 10
   comment: Reads account data from XP Github and upserts users into the Database
+- name: refreshPlayersFromComposeDB
+  webhook: '{{ACTION_BASE_ENDPOINT}}/composeDB/refreshPlayers'
+  schedule: 0 3 * * *
+  include_in_metadata: true
+  payload: {}
+  comment: Finds any cached player profiles more than four days old and queues them
+    to be recached from ComposeDB.
