--- packages/backend/src/handlers/actions/ceramic/routes.ts
@@ -1,8 +1,0 @@
-import express from 'express';
-
-import { asyncHandlerWrapper } from '../../../lib/apiHelpers.js';
-import linkProfileNode from './linkProfileNode.js';
-
-export const ceramicRoutes = express.Router();
-
-ceramicRoutes.post('/linkProfileNode', asyncHandlerWrapper(linkProfileNode));
