+++ packages/backend/src/handlers/actions/composeDB/updateSingleProfile/handler.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 
-import { queueRecache } from '../../../../lib/cacheHelper.js';
+import { queueRecache } from '../cacheHelper.js';
 
 export default async (req: Request, res: Response): Promise<void> => {
   const role = req.body.session_variables['x-hasura-role'];
