--- packages/backend/src/handlers/actions/idxCache/routes.ts
@@ -1,17 +1,0 @@
-import express from 'express';
-
-import { asyncHandlerWrapper } from '../../../lib/apiHelpers.js';
-import updateExpiredProfilesHandler from './updateExpiredProfiles/handler.js';
-import updateSingleProfileHandler from './updateSingleProfile/handler.js';
-
-export const cacheRoutes = express.Router();
-
-cacheRoutes.post(
-  '/updateSingle',
-  asyncHandlerWrapper(updateSingleProfileHandler),
-);
-
-cacheRoutes.post(
-  '/checkExpired',
-  asyncHandlerWrapper(updateExpiredProfilesHandler),
-);
