+++ packages/backend/src/handlers/auth-webhook/users.ts
@@ -1,7 +1,7 @@
 import Bottleneck from 'bottleneck';
 
-import { cacheProfile } from '../../lib/cacheHelper.js';
 import { client } from '../../lib/hasuraClient.js';
+import { cacheProfile } from '../actions/composeDB/cacheHelper.js';
 
 async function createPlayer(ethAddress: string, limiter: Bottleneck) {
   const { insert_profile: insert } = await client.CreatePlayerFromETH({
