+++ packages/backend/src/handlers/triggers/cacheComposeDBProfile.ts
@@ -1,11 +1,11 @@
 import Bottleneck from 'bottleneck';
 
 import { Player } from '../../lib/autogen/hasura-sdk.js';
-import { queueRecache } from '../../lib/cacheHelper.js';
+import { queueRecache } from '../actions/composeDB/cacheHelper.js';
 import { TriggerPayload } from './types.js';
 
 // This trigger is called when new accounts are created.
-export const cacheIDXProfile = async (
+export const cacheComposeDBProfile = async (
   payload: TriggerPayload<Player>,
   limiter: Bottleneck,
 ) => {
