+++ packages/web/graphql/mutations/idxCache.ts
@@ -1,5 +1,5 @@
 export const InsertCacheInvalidation = /* GraphQL */ `
   mutation InsertCacheInvalidation($playerId: uuid!) {
-    updateIDXProfile(playerId: $playerId)
+    updateCachedProfile(playerId: $playerId)
   }
 `;
