+++ packages/web/lib/hooks/ceramic/useSaveToComposeDB.ts
@@ -17,6 +17,7 @@ export const useSaveToComposeDB = () => {
 
   // When saving to ComposeDB, it's essential that we have the most recent value for user.ceramicProfileId so that we don't inadvertently create duplicate profile models. Thus, we specify 'network-only' to ALWAYS fetch the latest value of 'user' from Hasura.
   // Ideally, we should be invalidating the urql cache when persisting the ceramicProfileId below, but that would require switching over to Normalized caching (https://formidable.com/open-source/urql/docs/graphcache/normalized-caching/) which could be a big lift requiring a bunch of testing
+  // This is likely causing the warning "Cannot update a component while rendering a different component"
   const { user } = useUser({ requestPolicy: 'network-only' });
 
   const [, linkNode] = useLinkOwnCeramicNodeMutation();
